* upstream/master: (110 commits)
  Skip publisher flaky tests (#21657)
  backport: add 7.10 branch (#21635)
  [CI: Packaging] fix: push ubi8 images too (#21621)
  Docker build resiliance with a retry (#21587)
  Fix flaky FSWatch/FSScanner tests (#21625)
  chore: add versions 7.1x (#21670)
  [Elastic Agent] Reload fleet.kibana.hosts from policy change (#21599)
  Fix cyberark/corepas pipeline (#21643)
  Add openstack ssl provider in add_cloud_metadata (#21590)
  Add fips_enabled into all aws filesets (#21626)
  [Filebeat S3] Change log.file.path to be nested object (#21624)
  [CI] Change notification channel (#21559)
  Add `add_observer_metadata` `geo.name` to Quickstart (#21501)
  [E2E Tests] fix: set versions ony for PRs (#21608)
  Fix badger build in 386 (#21613)
  docs: Update timestamp.asciidoc (#20395)
  Remove nil-zero metrics and linux-exclusive metrics from Metricbeat (#21457)
  [Metricbeat] Use timestamp from CloudWatch for events (#21498)
  [Filebeat][S3 Input] Add support for FIPS endpoints (#21585)
  [Ingest Manager] Use new form of fleet API paths (#21478)
  ...

+++ x-pack/elastic-agent/pkg/artifact/download/http/verifier.go
@@ -59,9 +59,7 @@ func NewVerifier(config *artifact.Config, allowEmptyPgp bool, pgp []byte) (*Veri
 
 // Verify checks downloaded package on preconfigured
 // location agains a key stored on elastic.co website.
-func (v *Verifier) Verify(programName, version string) (bool, error) {
-	// TODO: think about verifying asc for prepacked beats
-
+func (v *Verifier) Verify(programName, version, artifactName string, removeOnFailure bool) (isMatch bool, err error) {
 	filename, err := artifact.GetArtifactName(programName, version, v.config.OS(), v.config.Arch())
 	if err != nil {
 		return false, errors.New(err, "retrieving package name")
@@ -72,16 +70,20 @@ func (v *Verifier) Verify(programName, version string) (bool, error) {
 		return false, errors.New(err, "retrieving package path")
 	}
 
-	isMatch, err := v.verifyHash(filename, fullPath)
-	if !isMatch || err != nil {
-		// remove bits so they can be redownloaded
-		os.Remove(fullPath)
-		os.Remove(fullPath + ".sha512")
-		os.Remove(fullPath + ".asc")
+	defer func() {
+		if removeOnFailure && (!isMatch || err != nil) {
+			// remove bits so they can be redownloaded
+			os.Remove(fullPath)
+			os.Remove(fullPath + ".sha512")
+			os.Remove(fullPath + ".asc")
+		}
+	}()
+
+	if isMatch, err := v.verifyHash(filename, fullPath); !isMatch || err != nil {
 		return isMatch, err
 	}
 
-	return v.verifyAsc(programName, version)
+	return v.verifyAsc(programName, version, artifactName)
 }
 
 func (v *Verifier) verifyHash(filename, fullPath string) (bool, error) {
@@ -127,7 +129,7 @@ func (v *Verifier) verifyHash(filename, fullPath string) (bool, error) {
 	return expectedHash == computedHash, nil
 }
 
-func (v *Verifier) verifyAsc(programName, version string) (bool, error) {
+func (v *Verifier) verifyAsc(programName, version, artifactName string) (bool, error) {
 	if len(v.pgpBytes) == 0 {
 		// no pgp available skip verification process
 		return true, nil
@@ -143,7 +145,7 @@ func (v *Verifier) verifyAsc(programName, version string) (bool, error) {
 		return false, errors.New(err, "retrieving package path")
 	}
 
-	ascURI, err := v.composeURI(programName, filename)
+	ascURI, err := v.composeURI(filename, artifactName)
 	if err != nil {
 		return false, errors.New(err, "composing URI for fetching asc file", errors.TypeNetwork)
 	}
@@ -177,7 +179,7 @@ func (v *Verifier) verifyAsc(programName, version string) (bool, error) {
 
 }
 
-func (v *Verifier) composeURI(programName, filename string) (string, error) {
+func (v *Verifier) composeURI(filename, artifactName string) (string, error) {
 	upstream := v.config.SourceURI
 	if !strings.HasPrefix(upstream, "http") && !strings.HasPrefix(upstream, "file") && !strings.HasPrefix(upstream, "/") {
 		// always default to https
@@ -190,7 +192,7 @@ func (v *Verifier) composeURI(programName, filename string) (string, error) {
 		return "", errors.New(err, "invalid upstream URI", errors.TypeNetwork, errors.M(errors.MetaKeyURI, upstream))
 	}
 
-	uri.Path = path.Join(uri.Path, "beats", programName, filename+ascSuffix)
+	uri.Path = path.Join(uri.Path, artifactName, filename+ascSuffix)
 	return uri.String(), nil
 }
